Start with backlog full of zeroes
The calculation of the backlog size using FLAG_BACKLOG_IS_PARTIAL and
FLAG_BACKLOG_IS_FILLED seems incorrect.  The number of bytes shifted
in save_backlog() is equal to match_scan_index, but st_step_search()
presumes that the number of bytes after the first shift is (window_length
- lookahead_sz) after the first shift and (window_length - 1) after
subsequent ones.

* The lines start = end - window_length + 1; in st_step_search() should
  drop the + 1.

* The number of bytes shifted the first time may be greater than
  (window_length - lookahead_size) if the last token pushed was a backref.

* When window length equals lookahead size, the formula (window_length -
  lookahead_size) gives us 0, when the number shifted obviously shouldn't
  be that. And the backlog is unlikely to be filled in two shifts.

Proposed solution:

Declare that the backlog is always full, initially full of zeroes. This
will make the backlog size calculation simple (equals window length!), and
is compatible with the decoder (all buffers are initially memset to 0).